refactor: better coding convention
