+++ examples/server/server.cc
@@ -1,4 +1,4 @@
-#include "cortex-common/EngineI.h"
+#include "cortex-common/enginei.h"
 #include "dylib.h"
 #include "httplib.h"
 #include "json/reader.h"
